extract ProjectType to a separate file
